VS2013 update.
git-svn-id: svn://svn.code.sf.net/p/hxcfloppyemu/code/HxCFloppyEmulator/HxCFloppyEmulator_software/trunk@1226 63fa2f70-6a9b-4bc4-b855-9c6a949b1d69

+++ sources/gui/cb_floppy_infos_window.cxx
@@ -455,7 +455,7 @@ int getdistance(int xsize,int ysize,int x_pos,int y_pos)
 {
 	int distance;
 
-	distance = (int)sqrt( pow( (xsize/2) - x_pos, 2 ) + pow( (ysize/2) - y_pos, 2 ) );
+	distance = (int)sqrt( pow( (float)((xsize/2) - x_pos), 2 ) + pow( (float)((ysize/2) - y_pos), 2 ) );
 
 	return distance;
 }
